moved load_promotion_status and resolve_model_bundle_uri_for_env from scripts.download_extracted_model_from_cloud to src.utils
